Redme + cleaning
