+++ HallucinationLab/Core/ExperimentModels.cs
@@ -3,6 +3,7 @@ namespace HallucinationLab.Core;
 public sealed class PromptCase
 {
     public string Id { get; set; } = string.Empty;
+    public string Truth { get; set; } = string.Empty;
     public string Prompt { get; set; } = string.Empty;
     public List<string> ExpectedFacts { get; set; } = new();
     public List<string> ForbiddenClaims { get; set; } = new();
