+++ HallucinationLab/Guard/IOutputGuard.cs
@@ -3,5 +3,5 @@ namespace HallucinationLab.Guard;
 public interface IOutputGuard
 {
     string Name { get; }
-    string Apply(string prompt, string rawOutput, IReadOnlyList<string> expectedFacts, IReadOnlyList<string> forbiddenClaims);
+    string Apply(string truth, string prompt, string rawOutput, IReadOnlyList<string> expectedFacts, IReadOnlyList<string> forbiddenClaims);
 }
