+++ README.md
@@ -13,46 +13,21 @@ Se agrego el proyecto `HallucinationLab` para comparar:
 1. Salida baseline de un modelo existente.
 2. La misma salida, postprocesada con un guard ontologico basado en este repositorio.
 
-El flujo permite medir si bajan las alucinaciones con dos listas por caso:
+### Resultados preliminares
 
-- `expectedFacts`: hechos que la respuesta deberia contener.
-- `forbiddenClaims`: afirmaciones consideradas alucinaciones.
+En una corrida local con `ReplayModelBackend` y 4 casos de prueba, el baseline quedo en una tasa media de alucinacion de 22.92% y el guard ontologico en 0%. En esos mismos casos, el guard redujo las afirmaciones prohibidas, pero a cambio aumento las abstenciones y dejo hechos esperados sin cubrir en dos de los cuatro escenarios.
 
-### Estructura principal
+Esto sugiere que, con estos parametros, el guard si puede cortar salida riesgosa, pero todavia no preserva bien la cobertura semantica cuando la respuesta original era util.
 
-- `HallucinationLab/Backends`: backend real (`OpenAiResponsesBackend`) y fallback (`ReplayModelBackend`).
-- `HallucinationLab/Guard`: guard de control (`PassThroughGuard`) y guard ontologico (`OntologiaOutputGuard`).
-- `HallucinationLab/Eval`: scoring y metricas agregadas.
-- `HallucinationLab/Samples`: casos y respuestas de ejemplo.
+### Evaluacion tecnica
 
-### Uso rapido
+Conviene seguir la linea de investigacion, pero no como sustituto directo de un motor real. El resultado preliminar es prometedor como mecanismo de contencion de alucinaciones, aunque hoy funciona mejor como capa de postprocesado conservadora que como solucion final.
 
-Ejecutar desde la raiz de la solucion:
+Para motores reales, la linea tiene sentido si se la continua en estas direcciones:
 
-```bash
-dotnet run --project HallucinationLab -- --cases HallucinationLab/Samples/cases.json --guard pass --out hallucination-report.json
-```
+1. Conectar la evaluacion a una referencia mas rica que simples coincidencias textuales.
+2. Medir el costo de abstenerse frente al beneficio de eliminar claims inventadas.
+3. Probar el guard con salidas reales de modelos, no solo con replay local.
+4. Ajustar el criterio para no penalizar respuestas correctas pero parciales.
 
-Para usar OpenAI real:
-
-```bash
-$env:OPENAI_API_KEY="tu_api_key"
-dotnet run --project HallucinationLab -- --cases HallucinationLab/Samples/cases.json --guard pass --model gpt-4o-mini --out hallucination-report.json
-```
-
-El comando genera:
-
-- Resumen por consola (baseline vs guarded).
-- Archivo `hallucination-report.json` con detalle por caso y metricas agregadas.
-
-### Como conectar tu modelo real
-
-1. Implementa `ITextModelBackend` en `HallucinationLab/Backends` para llamar tu proveedor (OpenAI, Azure, local, etc.).
-2. Mantene el mismo prompt para baseline y guarded.
-3. Usa el baseline como salida original del modelo.
-4. Aplica `OntologiaOutputGuard` para obtener la salida modificada.
-5. Compara metricas para estimar reduccion de alucinaciones.
-
-### Nota sobre OpenAI sin credenciales
-
-No existe un endpoint oficial de OpenAI para inferencia sin API key. Sin credenciales, el programa cae automaticamente a `ReplayModelBackend` para que el experimento pueda ejecutarse sin configuracion adicional.
+En resumen: si, vale la pena seguirla, pero como investigacion exploratoria para reducir alucinaciones y no como una garantia de calidad ya cerrada.
